requirements.txt
boto3==1.23.10
botocore==1.26.10
click==8.0.4
contextvars==2.4
dash==2.9.3
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-dangerously-set-inner-html==0.0.2
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses==0.8
Flask==2.0.3
immutables==0.19
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
numpy==1.19.5
packaging==21.3
pandas==1.1.5
plotly==5.14.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2023.3
s3transfer==0.5.2
six==1.16.0
tenacity==8.2.2
typing_extensions==4.1.1
urllib3==1.26.15
uWSGI==2.0.21
Werkzeug==2.0.3
xlrd==1.2.0
zipp==3.6.0
